涡旋检测模块
1. 删去sla和ow相关。2.新增合成速度。3.分离有无用api文件
