+++ static/js/scatter2y.js
@@ -14,9 +14,6 @@ function draw2yScatter() {
     ndxlatdim = ndx1d1dData.dimension(function (d) {
         return +d.lat;
     });
-    $("#reset-brush1").click(function(){
-        resetBrushed1();
-    });
     twoYScatterChart.margins({ top: 30, right: 30, bottom: 30, left: 30 })
         .shareTitle(false).elasticY(true)
         .legend(dc.legend().x(70).y(10).itemHeight(13).gap(5))
